requirements.txt
Flask~=1.1.2
pip~=20.2.4
setuptools~=50.3.2
python-dotenv~=0.15.0
flask-wtf

WTForms~=2.3.3
Jinja2~=2.11.2
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
MarkupSafe~=1.1.1

flask-sqlalchemy
flask-migrate
flask-login
email-validator
flask-mail
pyjwt
flask-bootstrap
flask-moment

elasticsearch
cryptography

psycopg2
gunicorn